+++ revo.js
@@ -1,22 +1,23 @@
 #!/usr/bin/env node
 
-var pkg = require('./package.json')
-  , fs = require('fs')
-  , _ = require('underscore')
-  , S = require('string')
-  , program = require('commander')
-  , yaml = require('js-yaml')
-  , scp = require('scp')
-  , container = require('./lib/container')
-  , application = require('./lib/application')
-  , component = require('./lib/component')
-  , util = require('./lib/util')
-  , repoManager = require('./lib/repo')
+var pkg = require('./package.json'),
+    debug = require('debug')('revo:'),
+    fs = require('fs'),
+    _ = require('underscore'),
+    S = require('string'),
+    program = require('commander'),
+    yaml = require('js-yaml'),
+    scp = require('scp'),
+    container = require('./lib/container'),
+    appService = require('./lib/services/appService'),
+    component = require('./lib/component'),
+    util = require('./lib/util'),
+    repoManager = require('./lib/repo');
 
 program
     .version(pkg.version)
     .usage('[command] [args] [options]')
-    // .option('-c, --components <components>', 'add <components> to an application. Use commas to separate the component names in the list')
+    .option('-c, --components <components>', 'add <components> to an application. Use commas to separate the component names in the list')
     // .option('-d, --data <files>', 'copy <files> to the data folder. Use commas to separate the file names in the list')
     .option('-f, --force', 'force on non-empty directory')
     // .option('-g, --config <config>', 'use components <config> file')
@@ -26,25 +27,14 @@ program
 program  
     .command('create [app_name]')
     .description('create a revo application')
-    .action(function(directory){
-        var destination;
-        if(program.recipe) {
-            var fullPathRecipe = S(program.recipe).ensureLeft('../').s;
-            try {
-                require.resolve(fullPathRecipe);
-                destination = require(fullPathRecipe).destination;
-            } catch(err) {
-                // recipe not found (ignore)
-            }
-        }
-        var path = destination || directory || ".";
-        util.emptyDirectory(path, function(empty){
-            if (empty || program.force) {
-                application.create({name:path, components:program.components, data:program.data, config:program.config, recipe:program.recipe});
-            } else {
-                util.abort('destination is not empty, aborting', 3);
-            }
-        });
+    .action(function(appName){
+        var opts = {
+            appName: appName,
+            recipeFile: program.recipe,
+            components: program.components,
+            force: !!program.force
+        };
+        appService.generateApp(opts);
     })
 
 program