+++ History.md
@@ -1,3 +1,8 @@
+0.1.1 - Apr 18, 2015
+====
+- use config for main configuration
+- dir layout restructuring & application service refactoring
+
 0.1.0 - Jan 15, 2015
 ====
 - initial version